PMM-7: New Test Tag

+++ tests/QAN/details_test.js
@@ -14,7 +14,7 @@ Before(async ({ I, queryAnalyticsPage }) => {
 });
 
 Scenario(
-  'Verify Details section tabs @qan',
+  'Verify Details section tabs @qan @gssapi-nightly',
   async ({
     I, queryAnalyticsPage,
   }) => {
@@ -28,7 +28,7 @@ Scenario(
 ).retry(1);
 
 Scenario(
-  'PMM-T223 - Verify time metrics are AVG per query (not per second) @qan',
+  'PMM-T223 - Verify time metrics are AVG per query (not per second) @qan @gssapi-nightly',
   async ({
     I, queryAnalyticsPage,
   }) => {
@@ -42,12 +42,19 @@ Scenario(
   },
 );
 
-const databaseEnvironments = [
-  // TODO: Unskip Percona Server when https://perconadev.atlassian.net/browse/PMM-13978 is finished.
-  // { dbType: 'PS', serviceName: 'ps_', queryTypes: ['SELECT', 'INSERT', 'DELETE', 'CREATE'] },
-  { serviceName: 'pdpgsql_pgsm_pmm_', queryTypes: ['SELECT s.first_name', 'INSERT INTO classes', 'DELETE FROM', 'CREATE TABLE classes '], cluster: '' },
-  { serviceName: 'rs101', queryTypes: ['db.students', 'db.runCommand', 'db.test'], cluster: 'replicaset' },
-];
+let databaseEnvironments;
+
+if (process.env.JOB_NAME.includes('gssapi')) {
+  databaseEnvironments = [
+    { serviceName: 'rs101_gssapi', queryTypes: ['db.students', 'db.runCommand', 'db.test'], cluster: 'replicaset' },
+  ];
+} else {
+  databaseEnvironments = [
+    { serviceName: 'ps_', queryTypes: ['SELECT s.first_name', 'INSERT INTO classes', 'DELETE FROM students', 'CREATE TABLE classes'], cluster: 'ps-single-dev-cluster' },
+    { serviceName: 'pdpgsql_pgsm_pmm_', queryTypes: ['SELECT s.first_name', 'INSERT INTO classes', 'DELETE FROM', 'CREATE TABLE classes '], cluster: '' },
+    { serviceName: 'rs101', queryTypes: ['db.students', 'db.runCommand', 'db.test'], cluster: 'replicaset' },
+  ];
+}
 
 Data(databaseEnvironments).Scenario(
   'PMM-T13 - Check Example, Explain, Plan and Table tabs for supported DBs @qan',
@@ -66,18 +73,23 @@ Data(databaseEnvironments).Scenario(
       await queryAnalyticsPage.data.verifyQueriesDisplayed(parameters);
       queryAnalyticsPage.data.selectRow(1);
       queryAnalyticsPage.waitForLoaded();
-      await queryAnalyticsPage.queryDetails.verifyExamples(parameters);
+      if (current.serviceName !== 'pgsql_pgss_pmm_') {
+        // pg stat statement does not support examples.
+        await queryAnalyticsPage.queryDetails.verifyExamples(parameters);
+      }
 
-      if (current.serviceName !== 'pdpgsql_pgsm_pmm_') {
+      if (!current.serviceName.includes('pgsql_') && !query.includes('CREATE')) {
+        // Explain is not for PostgreSQL and also not available for CREATE operations
         await queryAnalyticsPage.queryDetails.verifyExplain(parameters);
       }
 
-      if (current.serviceName === 'pdpgsql_pgsm_pmm_' && !query.includes('CREATE')) {
+      if (current.serviceName.includes('pgsql_') && !query.includes('CREATE')) {
+        // Tables are not available for PostgreSQL and also not available for CREATE operations
         await queryAnalyticsPage.queryDetails.verifyTables(parameters);
+      }
 
-        if (query.includes('SELECT')) {
-          await queryAnalyticsPage.queryDetails.verifyPlan(parameters);
-        }
+      if (current.serviceName === 'pdpgsql_pgsm_pmm_' && query.includes('SELECT')) {
+        await queryAnalyticsPage.queryDetails.verifyPlan(parameters);
       }
     }
   },
